providers: move get_instance_name() to providers.py
`get_instance_name()` is a rockcraft-specific function, so it is moved
to providers.py
