+++ rockcraft/providers/providers.py
@@ -17,6 +17,8 @@
 
 """Rockcraft-specific code to interface with craft-providers."""
 
+import pathlib
+
 from craft_providers import bases
 
 BASE_TO_BUILDD_IMAGE_ALIAS = {
@@ -24,3 +26,26 @@ BASE_TO_BUILDD_IMAGE_ALIAS = {
     "ubuntu:20.04": bases.BuilddBaseAlias.FOCAL,
     "ubuntu:22.04": bases.BuilddBaseAlias.JAMMY,
 }
+
+
+def get_instance_name(
+    *,
+    project_name: str,
+    project_path: pathlib.Path,
+) -> str:
+    """Formulate the name for an instance using each of the given parameters.
+
+    Incorporate each of the parameters into the name to come up with a
+    predictable naming schema that avoids name collisions across multiple
+    projects.
+
+    :param project_name: Name of the project.
+    :param project_path: Directory of the project.
+    """
+    return "-".join(
+        [
+            "rockcraft",
+            project_name,
+            str(project_path.stat().st_ino),
+        ]
+    )
