+++ tests/unit/providers/test_lxd.py
@@ -291,18 +291,6 @@ def test_get_command_environment_minimal(monkeypatch):
     }
 
 
-def test_get_instance_name(mock_path):
-    provider = providers.LXDProvider()
-
-    assert (
-        provider.get_instance_name(
-            project_name="my-project-name",
-            project_path=mock_path,
-        )
-        == "rockcraft-my-project-name-445566"
-    )
-
-
 @pytest.mark.parametrize("is_installed", [True, False])
 def test_is_provider_available(is_installed, mock_lxd_is_installed):
     mock_lxd_is_installed.return_value = is_installed
