+++ tests/unit/providers/test_provider.py
@@ -58,18 +58,6 @@ def test_get_command_environment_all_opts(provider_class, monkeypatch):
     }
 
 
-def test_get_instance_name(mock_path):
-    provider = providers.LXDProvider()
-
-    assert (
-        provider.get_instance_name(
-            project_name="my-project-name",
-            project_path=mock_path,
-        )
-        == "rockcraft-my-project-name-445566"
-    )
-
-
 @pytest.mark.parametrize(
     "name,expected_valid,expected_reason",
     [
