+++ tests/unit/providers/test_providers.py
@@ -1,0 +1,27 @@
+# -*- Mode:Python; indent-tabs-mode:nil; tab-width:4 -*-
+#
+# Copyright 2022 Canonical Ltd.
+#
+# This program is free software: you can redistribute it and/or modify
+# it under the terms of the GNU General Public License version 3 as
+# published by the Free Software Foundation.
+#
+# This program is distributed in the hope that it will be useful,
+# but WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+# GNU General Public License for more details.
+#
+# You should have received a copy of the GNU General Public License
+# along with this program.  If not, see <http://www.gnu.org/licenses/>.
+
+from rockcraft.providers import providers
+
+
+def test_get_instance_name(mock_path):
+    assert (
+        providers.get_instance_name(
+            project_name="my-project-name",
+            project_path=mock_path,
+        )
+        == "rockcraft-my-project-name-445566"
+    )
